Improved RB algorithm + added tests

+++ src/main/java/com/github/simple_tools/data/collection/Bag.java
@@ -174,7 +174,7 @@ public interface Bag<E>
     /**
      * @return The sum of the counts of the elements in the bag.
      */
-    int bagSize();
+    long bagSize();
     
     @Override
     default boolean isEmpty() {
