+++ src/main/java/com/github/simple_tools/data/collection/BagFunction.java
@@ -106,7 +106,7 @@ public class BagFunction<S, T>
     }
 
     @Override
-    public int bagSize() {
+    public long bagSize() {
         return getSrc().bagSize();
     }
